scraper/requirements.txt
requests
beautifulsoup4
pandas
supabase-py
python-dotenv
lxml 